TestEnvironment/Tripo/Tripo_Api_Automation/requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest
pyyaml
openpyxl
Appium-Python-Client

